various architectural improvements
