+++ apps/prairielearn/assets/scripts/esm-bundles/hydrated-components/AssessmentModulesPage.ts
@@ -1,5 +1,5 @@
 import { registerHydratedComponent } from '@prairielearn/react/hydrated-component';
 
-import { AssessmentModulesPage } from '../../../../src/pages/instructorCourseAdminModules/components/AssessmentModulesTable.js';
+import { AssessmentModulesPage } from '../../../../src/pages/instructorCourseAdminModules/components/AssessmentModulesPage.js';
 
 registerHydratedComponent(AssessmentModulesPage);
