+++ apps/prairielearn/assets/scripts/esm-bundles/hydrated-components/AssessmentSetsPage.ts
@@ -1,5 +1,5 @@
 import { registerHydratedComponent } from '@prairielearn/react/hydrated-component';
 
-import { AssessmentSetsPage } from '../../../../src/pages/instructorCourseAdminSets/components/AssessmentSetsTable.js';
+import { AssessmentSetsPage } from '../../../../src/pages/instructorCourseAdminSets/components/AssessmentSetsPage.js';
 
 registerHydratedComponent(AssessmentSetsPage);
