+++ apps/prairielearn/src/lib/assessment-modules.shared.ts
@@ -1,0 +1,6 @@
+/**
+ * The name of the implicit assessment module that sync always creates.
+ * Assessments without an explicit `module` belong to it, so it can never be
+ * renamed or deleted.
+ */
+export const DEFAULT_ASSESSMENT_MODULE_NAME = 'Default';
