+++ apps/prairielearn/src/lib/instructorFiles.ts
@@ -33,6 +33,34 @@ export interface InstructorFilePaths {
   }[];
 }
 
+/**
+ * The file-editor container for course-level files (e.g. `infoCourse.json`):
+ * the course root, excluding the directories owned by other editing contexts.
+ */
+export function getCourseContainer(coursePath: string) {
+  return {
+    rootPath: coursePath,
+    invalidRootPaths: [
+      path.join(coursePath, '.git'),
+      path.join(coursePath, 'questions'),
+      path.join(coursePath, 'courseInstances'),
+    ],
+  };
+}
+
+/**
+ * The file-editor container for course-instance-level files (e.g.
+ * `infoCourseInstance.json`): the course instance directory, excluding its
+ * assessments.
+ */
+export function getCourseInstanceContainer(coursePath: string, shortName: string) {
+  const rootPath = path.join(coursePath, 'courseInstances', shortName);
+  return {
+    rootPath,
+    invalidRootPaths: [path.join(rootPath, 'assessments')],
+  };
+}
+
 function getContextPaths(
   locals: UntypedResLocals,
 ): Pick<
@@ -49,12 +77,7 @@ function getContextPaths(
   if (locals.navPage === 'course_admin') {
     const rootPath = coursePath;
     return {
-      rootPath,
-      invalidRootPaths: [
-        path.join(rootPath, '.git'),
-        path.join(rootPath, 'questions'),
-        path.join(rootPath, 'courseInstances'),
-      ],
+      ...getCourseContainer(coursePath),
       cannotMove: [path.join(rootPath, 'infoCourse.json')],
       clientDir: path.join(rootPath, 'clientFilesCourse'),
       serverDir: path.join(rootPath, 'serverFilesCourse'),
@@ -63,8 +86,7 @@ function getContextPaths(
   } else if (locals.navPage === 'instance_admin') {
     const rootPath = path.join(coursePath, 'courseInstances', locals.course_instance.short_name);
     return {
-      rootPath,
-      invalidRootPaths: [path.join(rootPath, 'assessments')],
+      ...getCourseInstanceContainer(coursePath, locals.course_instance.short_name),
       cannotMove: [path.join(rootPath, 'infoCourseInstance.json')],
       clientDir: path.join(rootPath, 'clientFilesCourseInstance'),
       urlPrefix: `${locals.urlPrefix}/instance_admin`,
