+++ apps/prairielearn/src/models/assessment-module.ts
@@ -1,26 +1,14 @@
 import { z } from 'zod';
 
 import { loadSqlEquiv, queryRows } from '@prairielearn/postgres';
-import { IdSchema } from '@prairielearn/zod';
 
+import { AssessmentUsageSchema } from '../lib/client/assessment-usage.js';
 import { type AssessmentModule, AssessmentModuleSchema } from '../lib/db-types.js';
 
 const sql = loadSqlEquiv(import.meta.url);
 
-export const AssessmentForModuleSchema = z.object({
-  assessment_id: IdSchema,
-  tid: z.string(),
-  title: z.string(),
-  label: z.string(),
-  color: z.string(),
-  course_instance_id: IdSchema,
-  course_instance_short_name: z.string().nullable(),
-  course_instance_long_name: z.string().nullable(),
-});
-export type AssessmentForModule = z.infer<typeof AssessmentForModuleSchema>;
-
 export const AssessmentModuleWithAssessmentsSchema = AssessmentModuleSchema.extend({
-  assessments: z.array(AssessmentForModuleSchema),
+  assessments: z.array(AssessmentUsageSchema),
 });
 export type AssessmentModuleWithAssessments = z.infer<typeof AssessmentModuleWithAssessmentsSchema>;
